Updating the Introduction Section
I added missing information, cut out a few redudndant sentences and cleaned up the the grammar/spellings.

I removed the "Guidelines for the Teacher" section because those are specific to how we decide to structure the lessons and can vary. The guidelines will be part of the Curriculum, but not inside the book.